Moved entire State (except for compendium) to App to solve light switch bug. Removed two functions from LossCalc context and made them local  as they were only used once each.
